chore: move files to lib
