+++ lib/gauge-demo.gif
