+++ lib/has-color.js
