+++ lib/progress-bar.js
