+++ lib/render-template.js
