+++ lib/set-immediate.js
