+++ lib/set-interval.js
