+++ lib/spin.js
