+++ lib/template-item.js
