+++ lib/theme-set.js
