+++ lib/wide-truncate.js
